backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.8
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.1
fastar==0.8.0
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
newsapi-python==0.2.7
numpy==2.4.3
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.17.1
sentence-transformers==5.1.2
sentry-sdk==2.54.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==4.57.6
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
watchfiles==1.1.1
websockets==16.0
